Move file content helper into file resource
